New: generate sourcemap for Cantusdata source
Map cantus.min.js code back to its original sources. Therefore, copy all JS files across to Django so that browsers can display them for debugging.

As a side effect, this means that it is now possible to do a development build just by changing the production flag in the loading script, without any special-casing in the gulpfile.