WIP: use a collection for search results
This commit mostly implements the use of a collection for search results, which will ultimately simplify template rendering and allow for the removal of a level or two of views.

Things this commit temporarily breaks:

 - Volpiano results don't appear
 - Volpiano results would appear in the same row as the other results, if they did appear.
 - Manuscript names don't show up in the search result item views

A number of TODOs and FIXMEs have been added in the body of the code.