Revert "New: ChangeChant event parameter is now an object"
This reverts commit 301aff2afa6c446518dd6f17be854bac981adb3f.

The commit was intended to enable the passing of more information with
the ChangeChant event. However, the refactoring was never fully
implemented in the router, with the result that "chant=[Object object]"
would appear in the URL. Moreover, it is possible to pass multiple
parameters when triggering an event, so altering the first parameter
backwards incompatibly is unnecessary.

Refs #125