New: maintain a global registry of open chants per folio
Store the open chant for each folio and restore it when that folio is returned to. Currently this affects the URL but not the chant list in the user interface, which has other issues preventing reliable updates.

Getting this to work required some other patching to ensure that the global state is correctly being updated.

Refs #119