Fixed: expand the open chant when the chant list view is created
Previously the open chant would not be expanded in the chant list unless the user expanded the chant manually by clicking on the entry in the list.

This commit gets the chant list to synchronize itself with the global chant state. To do so, it changes the rendering and update logic in FolioView so that changes to the chant list don't cause the chant composite view to re-render.

Refs #119