Fixed: catch potential race condition in search loading
Previously, if a search result collection was set to fetch Solr results over multiple HTTP requests, and was then set to make a regular Backbone fetch before the Solr requests terminated, results from the earlier request could be added to the new collection even if it was supposed to have been reset.

This condition would never come up in the current code.