Fixed: Qualify "All fields" search input
Previously the default search input was "All fields." However, the Solr field it was linked to does index Volpiano, which is logically not a "text" field; you'll never enter a string expecting it to be found either in the Volpiano or in the chant text, etc.

This commit is a stopgap to try to correctly qualify expectations about the  scope of the field, pending the introduction of a clearer distinction between search types.

Refs #140