New: rework manuscript details presentation
Previously, the full manuscript title (a library identifier followed by a library-specific call number) was displayed in a heading element, with more metadata on the MS stored in a "Manuscript Details" tab.

This commit changes the heading format to the more compact "[provenance], [siglum]" and moves the manuscript details into a popover which is triggered by clicking on the heading.

Closes #199, refs #194