Fixed: make MEI CSV headings portable
Previously, the MEI import script relied on implementation-specific behaviour to get the order of columns when creating CSV files. This ordering varied between imports and (worse) could have varied for separate rows in the same import.

This commit makes the ordering of stable for each import, and tries to maintain the ordering between imports.