WIP: support searching for numerical chant modes
Previously, it was not possible to search by numerical chant modes because the search is realized using a Solr text field which does not index single characters.

This commit implements an client-side workaround which enables the desired behaviour.

It is insufficient in the long term both because this fix should really be applied to the Solr database and because the fix uses a hack to work around limitations in the client-side query representation.

Refs #211