Fixed: Content container can be larger than the window
Previously, the BrowserResizer set the height of the content container so that it only occupied the remaining visible area of the screen. This made it impossible to scroll on mobile devices. The restriction was also unnecessary, since any element whose height needs to be manually constrained has its height constrained separately.

This commit removes that restriction and is likely a fix for #200.

Refs #188