New: Move global resizing logic into specific views
Previously, the resizing of the Diva viewer and manuscript detail view was controlled globally by the BrowserResizer class.

This commit moves the resizing logic into the relevant views, with the effect that certain resizing that always used to happen now only happens when specific views are rendered, which will make it easier to make responsive styling the default.

The commit also removes some unnecessary event bindings for resizing. It shouldn't be necessary to trigger resizing on orientation change (because we do it on resize anyway) or on the rendering of some view in general. (We can add more specific callbacks if necessary.)

Refs #188