Refactor: Separate entry point for static pages
Previously, the static (i.e. flatpage-managed) parts of cantusdata loaded the same
scripts as the single page application components. This introduced two problems:
First, it required the loading of extra resources. Second, the application actually
tried to initialize the static pages as if they were part of the single page app.
This didn't have any user-visible effects because the process threw an error before
making unwanted changes, but it was non-obvious and potentially fragile.

This commit introduces an entry point for the static pages which omits many of
the dependencies of the Backbone app and uses initialization logic tailored for
the static pages, resolving both problems.