Fixed: Correctly test for Sass files
As a result of some kind of misleading documentation the condition which tested whether a source style file was a Sass file always returned true.

(The documentation says the condition should be  a "boolean or stat object or function that takes in a vinyl file and returns a boolean or RegularExpression that works on the file.path". Previously the condition was a function which returned a regular expression, which was always evaluated as a truthy value. Now it's just a regular expression.)