New: Show header content responsively
Previously, nav links were always made available via the side nav.
This commit makes them available directly in the header if the
screen is large enough. It also alters the behaviour of the page
title so that it only appears in the header if the current page is not
linked to by a nav link if the nav links are visible.