New: Load chant search results lazily
Previously, chant search results were loaded using a cascade of
requests which continued until all results were loaded. This
commit introduces a cutoff point after which search results are
only loaded after the user scrolls to the bottom of the screen. In
the process, it also introduces smarter logic for handling result
sorting, executing expensive sorting on the server rather than
the client.

Refs #195