New: Display OMR search results incrementally
Previously, all OMR search results displayed from the server
were displayed to the client immediately. This placed
considerable strain on the user agent, which had to render
thousands of result entries (with appropriate bindings), and on
the IIPImage server, which had to load potentially thousands of
result snippets which might never be seen.

This commit implements client-side lazy loading of results, similar to that for chant search. More results are displayed when the user scrolls down to the bottom the results.

Closes #195