New: Improve sizing of search type selector
Previously the width of the search type selector and input were
set as a percentage of the window width, which didn't work well
when the search region occupied a very large or small part of the
page.

This commit uses the flexbox sizing model to have the search type
selector it to its contents and have the search input occupy the
rest of the available width. On browsers without flexbox support
the search input has a fixed size.