Refactor: Set page height via CSS
Previously, the height of pages which were supposed to have
dimensions matching the viewport was handled via JS. This commit
reimplements that functionality in CSS using flexible box display.

This will still be hard to maintain because it requires sprinkling
`propagate-height` classes everywhere, and it's probably still kind
of buggy, but it should be better than it used to be, and it looks
less janky.

For user agents without flexbox support, absolute-value heights
are specified as a fallback.