New: Use URLs for manifests instead of hosting the files
A new data dumps file now contains the public manuscripts,
along with a URL pointing to their manifest, a URL to the
Cantus DB and a list of all the plugins they use.
To set those manuscripts to public in the Django database
and in Solr, use the new update_public_manuscripts
management command.
The folio mapping tool has also been updated to use URLs
for manifest files.

Refs #279