Modifies Solr synch signal to allow for concurrent db transactions
The prior instantiation of the signal would raise a set-related error
when multiple processes were making DB transations. Items would be added
to the SynchronizationSession objects _deletions and _additions attributes
during iteration (ie. when the execute method was called).

Adds an executing flag to the Solr synchronizer so that solr additions
and deletions will not be added to the set during iteration. Instead,
another SynchronizationSession will be created.