Add PENDING state to manuscript mapping
Adds a PENDING state to the manuscript model for is_mapped,
away from a binary state. This allows for the manuscript mapped
status to show as this interim state while processing on manuscript mapping
completes. This interim status shows on the admin/map_folios page
and the admin view during this time.