Update admin view when tasks are received
Previously, tasks were shown in the task status admin view
after a worker had started processing a task. This adds
a celery signal so that task status is updated when a task is
received so that tasks show up in the task status admin view
when they have been added to the queue, even if they haven't started
yet.