Set chants_imported = True after solr completes
Previously, a manuscript would have the chants_imported flag be set
to True even if the indexing of chants in solr failed. Since solr
indexing is an integral part of the chant import process, expected
behaviour would be that chants_imported is only set to True if both the
database transactions and solr indexing steps complete successfully.