Add genre to text field in solr schema
Multiple chant fields are added to the "text" element in the
solr schema so that those fields are included in an "All Text Fields"
search. Previously, the genre field was not included, so the genre
field was not searched in an "All Text Fields" search. This modifies
the solr schema so the genre field is included in that search.