Modify folio navigation search suggestions query
The solr query used to find suggestions for the folio navigation
search is modified to provide more useful results. We assume that
the string typed by the user is the beginning of the folio number
they wish to search (after some number of leading zeros) with a wildcard
following.

Additionally, this commit makes some changes to the folio_set.py
view requested by linting.