Fix error in constructing display query string
In #688, displayedQuery was introduced to modify volpiano query
strings returned from solr to remove escaped queries so the "expected"
query would be shown to the user in the SearchResultHeadingView.
This assumed that the query returned by solr was a string, but in the
case of a mode search the query returned is an array. This commit
now only removes the escape characters when the search is a volpiano
search (ie. only when necessary).