Highlight search modes on return to mode search
If a certain search field is returned to after a search in that
field has previously occured, the search results and the search
query are populated with the details of the previous search. Until
this commit, however, returning to the mode search view repopulated the
results of the search but did not re-select the options elements
that were searched.