Change handling of mode searches
All mode searches are matched against the mode_t_hidden field in
Solr. This will match single character searches (for mode searches
with queries "1", "2", etc but also special mode searches "Formulaic",
"Mode in Transposition", etc and simplifies the setSearchQueryOnBuilder
function. Searching only the full string (ie. removing the single-character
code) for special mode searches as a phrase means that special modes
are no longer matched when a single letter in the special mode name
is matched by the single-character code of the search (ie. previously,
the "T" of "T, Chant in Transposition" would match the "t" of
"Uncertain".

The default sorting of mode search results is adjusted to include
Solr's scoring so that when multiple options are selected in the
search, chants that meet multiple criteria are towards the top
of search results but any chant that meets one of the options
is still a possible result.

Comments are also added to the setSearchQueryOnBuilder function
to explain this reasoning.