Upgrade NodeJs to Node 12
In the process of this update, gulp had to be upgraded to v4. run-sequence
package was made irrelevant by the introduction of gulp.series in
gulp v4. Reorganization of the gulpfile.js was required due to changes
in task registration from gulp v3 to gulp v4 where gulp stopped allowing
forward references in task registration.